Knocked out some compiler warnings.

+++ isolde/src/openmm/custom_forces.cpp
@@ -120,8 +120,8 @@ customexternalforce_add_particles(void *force, size_t n, int *particle_indices, 
     try {
         int n_params = f->getNumPerParticleParameters();
         std::vector<double> param_vec(n_params);
-        for (size_t i=0; i<n; ++i) {
-            for (size_t j=0; j<n_params; ++j)
+        for (int i=0; i<n; ++i) {
+            for (int j=0; j<n_params; ++j)
                 param_vec[j] = *params++;
             *(force_indices++) = f->addParticle(*(particle_indices++), param_vec);
         }
